Break dependency on ContentIndexParser in FlowParser and SurveyParser
